Fix: follow course guideline for decryption rounds
The decryption is done through multiple rounds, most of which has the
same steps, except the first and the last rounds which has different
number of steps. This division of where to start and end the round could
be done in a number of ways. The way I coded previously was taken from a
Github project, but it did not follow the guideline advised by my course
teacher. Since this is an assignment, I decided to tweak it lest I get
discredited for plagiarism.